Split card data structs into separate files
