--- Assets/Scripts/Data/Structs/CardStructs.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: eb7c27027a1cf1642bc14f3d24c8010a
