+++ Assets/Scripts/Data/Structs/CriticalData.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 921ca1303c29fceba489710728fa0590
