+++ Assets/Scripts/Data/Structs/GrowthRateData.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 6f99eb9435a6799449dd469ae8ed52fb
