+++ Assets/Scripts/Data/Structs/UniqueSkillData.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 5ed2c4054bc024ee779f02e6cf06c4c7
